requirements.txt
asgiref==3.7.2
Django==5.0.3
django-ckeditor==6.7.1
django-extensions==3.2.3
django-js-asset==2.2.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
phonenumbers==8.13.31
pillow==10.2.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
